Ensure GammaLuts are computed at compile time, misc code fixes - Removed unnecessary parentheses. - Changed GammaLut values to u8. - Fixed gbarunner9.ld indentation for consistency.

+++ code/core/arm9/source/GbaColorCorrection/ColorLut.h
@@ -5,5 +5,5 @@
 
 extern u16 gColorLut[COLOR_LUT_SIZE];
 
-void clut_initColorCorrection(const ColorProfile* preset, int gammaIndex);
+void clut_initColorCorrection(const ColorProfile* preset, u8 gammaIndex);
 void clut_disableColorCorrection();
